// src/app/Interfaces/ProductCard.ts
export interface ProductCard {
  product_id: string;
  product_id_v2: string;
  product_title: string;
  product_description: string;
  product_photo: string;
  product_rating: number;
  product_page_url: string;
  product_num_reviews: number;
  typical_price_range: string;
  price: string;
  original_price:string;
}